Adds `size_t` casting to `int` vars to fix warning about unsigned to
signed comparisons, removes an unused `auto` var.

These fixes were needed to allow `make test-dyno` to proceed without
failure - unsure how the original PRs made it passed the dyno-test CI
stage.

TESTING:

- [x] local `make test-dyno` works again

[reviewed by @jabraham17 - thanks!]

+++ frontend/test/parsing/testParseBegin.cpp
@@ -136,7 +136,7 @@ static void test3(Parser* parser) {
       "begin with () { }\n"
       "begin with ref A { }\n");
   auto numErrors = 5;
-  assert(guard.errors().size() == numErrors);
+  assert(guard.errors().size() == (size_t) numErrors);
   assert("invalid intent expression in 'with' clause" == guard.error(1)->message());
   assert("'with' clause cannot be empty" == guard.error(2)->message());
   assert("missing parentheses around 'with' clause intents" == guard.error(4)->message());
